+++ test/endpoint.js
@@ -168,12 +168,6 @@ describe('stemmers', function() {
   });
 });
 
-describe('structures', function() {
-  require('./structures/bk-tree.js');
-  require('./structures/symspell.js');
-  require('./structures/vp-tree.js');
-});
-
 describe('tag', function() {
   require('./tag/averaged-perceptron.js');
 });